Switch to precision='16-mixed' (new API) to avoid AMP+manual_opt deadlock
precision=16 is deprecated in PL — the legacy code path's gradient scaler
can desync across ranks when one rank's loss becomes inf/nan and skips
opt.step(), causing the next NCCL collective to hang.

+++ train.py
@@ -422,7 +422,7 @@ def cli_main():
         logger=False,
         num_sanity_val_steps=0,
         sync_batchnorm=True,
-        precision=16,
+        precision="16-mixed",
         callbacks=[checkpoint_callback],
         default_root_dir=config['save_dir'],
         reload_dataloaders_every_n_epochs=1,
